fix: switch back from ty to pyrefly, pre-commit to prek
they weren't kidding that ty is pre-alpha and not ready, theres so many
things it doesn't catch that every other checker does (see
[here](https://github.com/astral-sh/ty/issues/1141#issuecomment-3262811695))

pyrefly seems to work fine on the project in its current state now
(though no news on [my bug
report](https://github.com/facebook/pyrefly/issues/1049))
and i really don't want to use pyright or mypy, so i think pyrefly it is
and will be, until ty hits beta at least.

additionally, swap out pre-commit for prek. i like the premise of
pre-commit but the author is a bit of an ass so i am glad to see a shiny
rust🚀⭐🚀 replacement.